Backend/requirements.txt
# Core FastAPI and async support
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LangChain and OpenAI
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28
openai==1.59.6

# MongoDB
pymongo==4.10.1
motor==3.6.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Google OAuth
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Environment and utilities
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1

# Email support
aiosmtplib==3.0.2
email-validator==2.2.0

# CORS
python-cors==1.0.0

# Rate limiting
slowapi==0.1.9

# Logging
python-json-logger==3.2.1
